correction de la fonction de collision

+++ util/keywords.h
@@ -9,7 +9,7 @@
 #define KEYWORDS_H_
 
 #ifndef M_PI
-#define M_PI 3.14159265358979323846
+#define M_PI ((double)3.14159265358979323846)
 #endif
 
 enum Couleur {
